Created Todo List App
Created a todo list app using HTML, CSS, and JavaScript.

Todo list is saved to localStorage in order to create a persistent todo list when closing/opening the page.

Users are able to dynamically add and remove items from the todo list. Users are also able to mark todo list items as complete.